refactor: 역할에 따른 파일 분리
